requirements.txt
azure-ai-agents
azure-identity
python-dotenv
